requirements.txt
# requirements.txt
jinja2
unidecode
qrcode
python-dotenv
fastapi
ezodf
pydantic
pandas
odfpy
uvicorn[standard]
